+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # v3.0.2
 * Fixed: Elasticsearch5 throwing unsupported operation exception when adding/updating a property
+* Added: Added a has method to the Query class to allow searches on all properties of a particular data type.
+* Added: Added a has method to the Query class to allow searches for a value across multiple properties.
+* Added: Implemented support all compare operators for DateOnly field types when using Elasticsearch 5.
 
 # v3.0.1
 * Changed: Find Path to not return a path that contains one or more vertices that is can't be retrieved because of visibility restrictions